requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
antlr-python-runtime=4.9.3=pyhd8ed1ab_1
asttokens=2.0.5=pyhd3eb1b0_0
attrs=22.2.0=pyh71513ae_0
backcall=0.2.0=pyhd3eb1b0_0
brotlipy=0.7.0=py310h5764c6d_1005
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.12.7=ha878542_0
certifi=2022.12.7=py310h06a4308_0
cffi=1.15.1=py310h255011f_3
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
coverage=7.0.1=py310h1fa729e_0
cryptography=39.0.0=py310h65dfdc0_0
cudatoolkit=10.1.243=h6bb024c_0
debugpy=1.5.1=py310h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
dgl=0.9.1=pypi_0
entrypoints=0.4=py310h06a4308_0
exceptiongroup=1.1.0=pyhd8ed1ab_0
executing=0.8.3=pyhd3eb1b0_0
fastcluster=1.2.6=pypi_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h7b6447c_0
googledrivedownloader=0.4=pyhd3deb0d_1
html5lib=1.1=pyh9f0ad1d_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.0.0=pyha770c72_0
iniconfig=1.1.1=pyh9f0ad1d_0
intel-openmp=2022.1.0=h9e868ea_3769
ipykernel=6.9.1=py310h06a4308_0
ipython=8.4.0=py310h06a4308_0
isodate=0.6.1=pyhd8ed1ab_0
jedi=0.18.1=py310h06a4308_1
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=h7f8727e_0
jupyter_client=7.2.2=py310h06a4308_0
jupyter_core=4.10.0=py310h06a4308_0
lapjv=1.3.24=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libblas=3.9.0=16_linux64_openblas
libcblas=3.9.0=16_linux64_openblas
libdeflate=1.8=h7f8727e_5
libffi=3.4.2=h6a678d5_6
libgcc-ng=12.2.0=h65d4601_19
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
liblapack=3.9.0=16_linux64_openblas
libopenblas=0.3.21=pthreads_h78a6416_3
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.1=h4ff587b_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=12.2.0=h46fd767_19
libtiff=4.4.0=hecacb30_2
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py310h5764c6d_2
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl=2022.1.0=hc2b9512_224
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py310h06a4308_0
networkx=2.8.8=pyhd8ed1ab_0
ninja=1.11.0=h924138e_0
nltk=3.7=pypi_0
numpy=1.23.5=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
omegaconf=2.3.0=pyhd8ed1ab_0
openssl=1.1.1s=h0b41bf4_1
packaging=22.0=pyhd8ed1ab_0
pandas=1.5.2=py310h769672d_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.3.0=py310hace64e9_1
pip=22.2.2=py310h06a4308_0
pluggy=1.0.0=pyhd8ed1ab_5
prompt-toolkit=3.0.20=pyhd3eb1b0_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=23.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
pytest=7.2.0=pyhd8ed1ab_2
pytest-cov=4.0.0=pyhd8ed1ab_0
python=3.10.8=h7a1cb2a_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytorch=1.12.1=cpu_py310h75c9ab6_0
pytorch_scatter=2.1.0=cpu_py310h10b8b27_1
pytz=2022.7=pyhd8ed1ab_0
pyyaml=6.0=py310h5764c6d_5
pyzmq=23.2.0=py310h6a678d5_0
rdflib=6.2.0=pyhd8ed1ab_0
readline=8.2=h5eee18b_0
regex=2022.10.31=pypi_0
requests=2.28.1=pyhd8ed1ab_1
scikit-learn=1.2.0=py310h209a8ca_0
scipy=1.9.3=pypi_0
setuptools=65.5.0=py310h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0.post1=pypi_0
sleef=3.5.1=h9b69904_2
sqlite=3.40.0=h5082296_0
stack_data=0.2.0=pyhd3eb1b0_0
texttable=1.6.7=pypi_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd8ed1ab_0
tomli=2.0.1=pyhd8ed1ab_0
torch=1.13.0=pypi_0
torch-geometric=2.2.0=pypi_0
torch-sparse=0.6.16=pypi_0
torchmetrics=0.11.0=pypi_0
torchvision=0.2.2=py_3
tornado=6.1=py310h7f8727e_0
tqdm=4.64.1=pyhd8ed1ab_0
traitlets=5.1.1=pyhd3eb1b0_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022f=h04d1e81_0
urllib3=1.26.13=pyhd8ed1ab_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py_1
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=h5eee18b_0
yaml=0.2.5=h7f98852_2
zeromq=4.3.4=h2531618_0
zipp=3.11.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
